+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Include advanced tactics in calculated attack and defense bonus
 * Show tooltips for resources in plunder screen in local (server) language ([#12](https://github.com/veger/foei/issues/12))
 * Use localized names for the units ([#12](https://github.com/veger/foei/issues/12))
+* Show unit images instead of unit names
 
 # 0.5.0
 * The Crow's Nest cannot be plundered